test: 행사 조회 테스트
