+++ service/main-server/src/main/java/org/codeNbug/mainserver/MainServerApplication.java
@@ -4,10 +4,8 @@ import org.codenbug.user.security.config.CorsProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
-@EnableJpaAuditing
 @EnableConfigurationProperties(CorsProperties.class) // CORS 설정을 위한 프로퍼티 클래스 활성화 (임시)
 public class MainServerApplication {
