+++ service/main-server/src/main/java/org/codeNbug/mainserver/global/config/AppConfig.java
@@ -2,10 +2,12 @@ package org.codeNbug.mainserver.global.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
 @EnableScheduling
+@EnableJpaAuditing
 @ComponentScan(basePackages = {"org.codeNbug.mainserver", "org.codenbug.user",
 	"org.codenbug.common"})
 public class AppConfig {
